Simplifying persistence class names
